server/oauth2/oauth2_connect.go
// Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
// See License for license information.

package oauth2

import (
	"net/http"

	"github.com/mattermost/mattermost-plugin-mscalendar/server/mscalendar"
)

func oauth2Connect(w http.ResponseWriter, r *http.Request) {
	userID := r.Header.Get("Mattermost-User-ID")
	if userID == "" {
		http.Error(w, "Not authorized", http.StatusUnauthorized)
		return
	}

	mscalendar := mscalendar.FromContext(r.Context())
	url, err := mscalendar.InitOAuth2(userID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	http.Redirect(w, r, url, http.StatusFound)
}
